backend/requirements.txt
fastapi
uvicorn
requests 
python-multipart
httpx
python-dotenv  # Para gestionar variables de entorno